[FEATURE] #97 : Analysis 관련 model 및 dto 구현
