+++ android/core/model/src/main/java/com/jaino/model/analysis/AnalysisHistory.kt
@@ -1,0 +1,6 @@
+package com.jaino.model.analysis
+
+data class AnalysisHistory (
+    val id : Long,
+    val date: String
+)
