+++ android/core/model/src/main/java/com/jaino/model/analysis/AnalysisId.kt
@@ -1,0 +1,5 @@
+package com.jaino.model.analysis
+
+data class AnalysisId(
+    val analysisId: Long
+)
