--- android/core/model/src/main/java/com/jaino/model/user/Analysis.kt
@@ -1,6 +1,0 @@
-package com.jaino.model.user
-
-data class Analysis (
-    val id : Long,
-    val date: String
-    )
