+++ android/core/network/src/main/java/com/jaino/network/model/response/analysis/AnalysisHistoryResponse.kt
@@ -1,0 +1,15 @@
+package com.jaino.network.model.response.analysis
+
+import com.jaino.model.analysis.AnalysisHistory
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class AnalysisHistoryResponse (
+    val id: Long,
+    val date: String
+){
+    fun toAnalyzeHistory() : AnalysisHistory = AnalysisHistory(
+        id = id,
+        date = date
+    )
+}
