+++ android/core/network/src/main/java/com/jaino/network/model/response/analysis/AnalysisIdResponse.kt
@@ -1,0 +1,12 @@
+package com.jaino.network.model.response.analysis
+
+import com.jaino.model.analysis.AnalysisId
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class AnalysisIdResponse(
+    @SerialName("analysis-id") val analysisId : Long
+){
+    fun toAnalysisId(): AnalysisId = AnalysisId(analysisId)
+}
