--- android/core/network/src/main/java/com/jaino/network/model/response/user/AnalysisResponse.kt
@@ -1,15 +1,0 @@
-package com.jaino.network.model.response.user
-
-import com.jaino.model.user.Analysis
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class AnalysisResponse (
-    val id: Long,
-    val date: String
-){
-    fun toAnalyze() : Analysis = Analysis(
-        id = id,
-        date = date
-    )
-}
